Resolve S12P11A609-39

See merge request s12-webmobile2-sub1/S12P11A609!28

+++ backend/src/main/java/com/ssafy/d3v/backend/member/entity/Member.java
@@ -48,10 +48,10 @@ public class Member extends BaseEntity {
     private String githubUrl;
 
     @Column(name = "max_streak")
-    private Integer maxStreak;
+    private Long maxStreak;
 
     @Column(name = "ongoing_streak")
-    private Integer ongoingStreak;
+    private Long ongoingStreak;
 
     @Column(name = "provider_type")
     private String providerType;
